+++ notifications_utils/recipient_validation/phone_number.py
@@ -44,7 +44,7 @@ international_phone_info = namedtuple(
         "international",
         "crown_dependency",
         "country_prefix",
-        "billable_units",
+        "rate_multiplier",
     ],
 )
 
@@ -230,17 +230,38 @@ class PhoneNumber:
 
     def get_international_phone_info(self):
         return international_phone_info(
-            international=phonenumbers.region_code_for_number(self.number) != "GB",
-            crown_dependency=self._is_a_crown_dependency_number(),
+            international=self.is_international_number(),
+            crown_dependency=self.is_a_crown_dependency_number(),
             country_prefix=self.prefix,
-            billable_units=INTERNATIONAL_BILLING_RATES[self.prefix]["billable_units"],
+            rate_multiplier=INTERNATIONAL_BILLING_RATES[self.prefix]["rate_multiplier"],
         )
 
-    def _is_a_crown_dependency_number(self):
+    def is_international_number(self):
+        """
+        Returns True for phone numbers that either have a GB country code
+        or that are OFCOM TV numbers. libphonenumber only contains actually
+        valid phonenumbers in its metadata, so TV numbers will return null
+        values calling methods like `phonenumbers.region_code_fornumber` so
+        need handling as a special case.
+        """
+        if phonenumbers.region_code_for_number(self.number) == "GB":
+            return False
+        elif self._is_tv_number(self.number):
+            return False
+        else:
+            return True
+
+    def is_a_crown_dependency_number(self):
         """
         Returns True for phone numbers from Jersey, Guernsey, Isle of Man, etc
+        TV numbers are an edge case where libphonenumber cannot accurately
+        handle them (they're not actually valid numbers). In that case we
+        always want to return False as we consider them to be UK numbers.
         """
-        return self.is_uk_phone_number() and phonenumbers.region_code_for_number(self.number) != "GB"
+        if self._is_tv_number(self.number):
+            return False
+        else:
+            return self.is_uk_phone_number() and phonenumbers.region_code_for_number(self.number) != "GB"
 
     def should_use_numeric_sender(self):
         """